requirements.txt
asgiref==3.2.3
Django==3.0.2
django-cors-headers==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
numpy==1.18.1
Pillow==7.0.0
pytz==2019.3
scipy==1.4.1
sqlparse==0.3.0
